Reorganized source code, changed Home and Create fragments to use InjectorUtils with factory refs
