+++ app/src/main/AndroidManifest.xml
@@ -9,16 +9,16 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".CounterActivity"></activity>
+        <activity android:name=".views.CounterActivity"></activity>
         <activity
-            android:name=".CreateActivity"
-            android:parentActivityName=".MainActivity"/>
-        <activity android:name=".DisplayMessageActivity">
+            android:name=".views.CreateActivity"
+            android:parentActivityName=".views.MainActivity"/>
+        <activity android:name=".views.DisplayMessageActivity">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value=".MainActivity" />
+                android:value=".views.MainActivity" />
         </activity>
-        <activity android:name=".MainActivity" android:label="@string/app_name">
+        <activity android:name=".views.MainActivity" android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
