+++ app/src/main/java/com/example/myfirstapp/fragments/Home.kt
@@ -1,18 +1,19 @@
-package com.example.myfirstapp
+package com.example.myfirstapp.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.fragment.app.activityViewModels
+import androidx.fragment.app.viewModels
 import androidx.navigation.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.example.myfirstapp.R
+import com.example.myfirstapp.TaskAdapter
 import com.example.myfirstapp.models.Task
-import com.example.myfirstapp.services.AppDatabase
+import com.example.myfirstapp.utilities.InjectorUtils
 import com.example.myfirstapp.viewmodels.TaskViewModel
-import com.example.myfirstapp.viewmodels.TaskViewModelFactory
 import kotlinx.android.synthetic.main.fragment_home.view.*
 
 // TODO: Rename parameter arguments, choose names that match
@@ -30,10 +31,10 @@ class Home : Fragment() {
     private var param1: String? = null
     private var param2: String? = null
 
-    val application = requireActivity().application
-    //private val model: TaskViewModel by activityViewModels { TaskViewModelFactory(application, AppDatabase.getDatabase(application).taskDao()) }
+    val context = requireActivity()
+    private val model: TaskViewModel by viewModels { InjectorUtils.provideTaskViewModelFactory(context) }
 
-    private val model: TaskViewModel by activityViewModels()
+    //private val model: TaskViewModel by activityViewModels()
 
     // Recycler View components
     private lateinit var recyclerView: RecyclerView
