+++ app/src/main/java/com/example/myfirstapp/viewmodels/TaskViewModel.kt
@@ -1,20 +1,14 @@
 package com.example.myfirstapp.viewmodels
 
-import android.app.Application
-import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
-import com.example.myfirstapp.services.AppDatabase
 import com.example.myfirstapp.models.Task
-import com.example.myfirstapp.services.DBService
+import com.example.myfirstapp.data.DBService
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
 
 class TaskViewModel(private val dbService: DBService<Task>) : ViewModel() {
-
-    //private val dbService: DBService<Task> =  AppDatabase.getDatabase(application).taskDao()
 
     /**
      * Launching a new coroutine to interact w/ data
@@ -22,13 +16,5 @@ class TaskViewModel(private val dbService: DBService<Task>) : ViewModel() {
     fun getTasks(): Array<Task> = runBlocking { dbService.getAll() }
     fun insert(task: Task) = viewModelScope.launch(Dispatchers.IO) { dbService.insert(task) }
     fun delete(task: Task) = viewModelScope.launch(Dispatchers.IO) { dbService.delete(task) }
-
-}
-
-class TaskViewModelFactory(private val application: Application, private val dbService: DBService<Task>) : ViewModelProvider.AndroidViewModelFactory(application) {
-
-//    override fun <T : ViewModel?> create(modelClass: Class<T>): T {
-//        return TaskViewModel(application, dbService) as T
-//    }
 
 }
