+++ app/src/main/java/com/example/myfirstapp/viewmodels/TaskViewModelFactory.kt
@@ -1,10 +1,9 @@
-package com.example.myfirstapp.utilities
+package com.example.myfirstapp.viewmodels
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.example.myfirstapp.models.Task
-import com.example.myfirstapp.services.DBService
-import com.example.myfirstapp.viewmodels.TaskViewModel
+import com.example.myfirstapp.data.DBService
 
 class TaskViewModelFactory(private val repo: DBService<Task>): ViewModelProvider.Factory {
