+++ app/src/main/java/com/example/myfirstapp/views/MainActivity.kt
@@ -1,16 +1,13 @@
-package com.example.myfirstapp
+package com.example.myfirstapp.views
 
-import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.view.MenuItem
-import android.view.View
-import android.widget.EditText
 import androidx.appcompat.widget.Toolbar
 import androidx.databinding.DataBindingUtil
 import androidx.navigation.findNavController
 import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.setupWithNavController
+import com.example.myfirstapp.R
 import com.example.myfirstapp.databinding.ActivityMainBinding
 
 const val EXTRA_MESSAGE = "com.example.myfirstapp.MESSAGE"
@@ -21,7 +18,9 @@ class MainActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        binding = DataBindingUtil.setContentView(this, R.layout.activity_main)
+        binding = DataBindingUtil.setContentView(this,
+            R.layout.activity_main
+        )
 
         setSupportActionBar(binding.toolbar)
         val navController = findNavController(R.id.nav_host_fragment)
