+++ app/src/main/res/layout/activity_counter.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CounterActivity">
+    tools:context=".views.CounterActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
