+++ app/src/main/res/layout/activity_display_message.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DisplayMessageActivity">
+    tools:context=".views.DisplayMessageActivity">
 
     <TextView
         android:id="@+id/textView"
