+++ app/src/main/res/layout/fragment_create.xml
@@ -5,7 +5,7 @@
     <FrameLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".Create">
+        tools:context=".fragments.Create">
 
         <ScrollView
             android:layout_width="match_parent"
